Fix a few minor issues

+++ src/core/modules/Nebula.hpp
@@ -350,8 +350,8 @@ private:
 	static StelTextureSP texRegion;				// The symbolic dashed shape texture
 	static StelTextureSP texPointElement;
 	static StelTextureSP texPlanetaryNebula;		// Type 3
-	static double hintsBrightness;
-	static double labelsBrightness;
+	static float hintsBrightness;
+	static float labelsBrightness;
 
 	static Vec3f labelColor;				// The color of labels
 	static QMap<Nebula::NebulaType, Vec3f>hintColorMap;	// map for rapid lookup. Updated by NebulaMgr whenever a color changes.
